Updated to version used in submission
Updated this to the version used in the submitted manuscript. Data collection methodology was changed to improve results and most changes reflect the new data parameters. Removed the check on whether a "shoreline" version crosses itself at a given longitude. This may be worth going back to at some point, but does not majorly affect the results (as sections where this is important are removed from the stats in the paper, e.g., where "shorelines" track near-longitudinally, with explanation).